+++ apps/synt-vedtakshistorikk-service/src/main/java/no/nav/testnav/apps/syntvedtakshistorikkservice/consumer/command/search/PersonSearchCommand.java
@@ -1,4 +1,4 @@
-package no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.command;
+package no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.command.search;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
